refactor: move and restructure config-related enums into a dedicated `Config` namespace for improved organization
