+++ src/Extensions/NeoServer.Scripts.LuaJIT/ConfigManager.cs
@@ -1,5 +1,6 @@
 ﻿using LuaNET;
 using NeoServer.Scripts.LuaJIT.Enums;
+using NeoServer.Scripts.LuaJIT.Enums.Config;
 using NeoServer.Scripts.LuaJIT.Interfaces;
 using Serilog;
