+++ src/Extensions/NeoServer.Scripts.LuaJIT/ConfigurationMap.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.Configuration;
 using NeoServer.Scripts.LuaJIT.Enums;
+using NeoServer.Scripts.LuaJIT.Enums.Config;
 
 namespace NeoServer.Scripts.LuaJIT;
