+++ src/Extensions/NeoServer.Scripts.LuaJIT/Enums/Config/BooleanConfigType.cs
@@ -1,0 +1,86 @@
+namespace NeoServer.Scripts.LuaJIT.Enums.Config;
+
+public enum BooleanConfigType
+{
+    ALLOW_CHANGEOUTFIT,
+    ONE_PLAYER_ON_ACCOUNT,
+    AIMBOT_HOTKEY_ENABLED,
+    REMOVE_RUNE_CHARGES,
+    EXPERIENCE_FROM_PLAYERS,
+    FREE_PREMIUM,
+    REPLACE_KICK_ON_LOGIN,
+    BIND_ONLY_GLOBAL_ADDRESS,
+    OPTIMIZE_DATABASE,
+    MARKET_PREMIUM,
+    EMOTE_SPELLS,
+    STAMINA_SYSTEM,
+    WARN_UNSAFE_SCRIPTS,
+    CONVERT_UNSAFE_SCRIPTS,
+    CLASSIC_ATTACK_SPEED,
+    SCRIPTS_CONSOLE_LOGS,
+    REMOVE_WEAPON_AMMO,
+    REMOVE_WEAPON_CHARGES,
+    REMOVE_POTION_CHARGES,
+    GLOBAL_SERVER_SAVE_NOTIFY_MESSAGE,
+    GLOBAL_SERVER_SAVE_CLEAN_MAP,
+    GLOBAL_SERVER_SAVE_CLOSE,
+    GLOBAL_SERVER_SAVE_SHUTDOWN,
+    FORCE_MONSTERTYPE_LOAD,
+    HOUSE_OWNED_BY_ACCOUNT,
+    CLEAN_PROTECTION_ZONES,
+    ALLOW_BLOCK_SPAWN,
+    HOUSE_PURSHASED_SHOW_PRICE,
+    ONLY_INVITED_CAN_MOVE_HOUSE_ITEMS,
+    WEATHER_RAIN,
+    WEATHER_THUNDER,
+    TOGGLE_FREE_QUEST,
+    ONLY_PREMIUM_ACCOUNT,
+    TOGGLE_MAP_CUSTOM,
+    ALL_CONSOLE_LOG,
+    STAMINA_TRAINER,
+    STAMINA_PZ,
+    PUSH_WHEN_ATTACKING,
+    SORT_LOOT_BY_CHANCE,
+    TOGGLE_SAVE_INTERVAL,
+    TOGGLE_SAVE_INTERVAL_CLEAN_MAP,
+    PREY_ENABLED,
+    PREY_FREE_THIRD_SLOT,
+    TASK_HUNTING_ENABLED,
+    TASK_HUNTING_FREE_THIRD_SLOT,
+    STASH_MOVING,
+    TOGGLE_IMBUEMENT_SHRINE_STORAGE,
+    TOGGLE_IMBUEMENT_NON_AGGRESSIVE_FIGHT_ONLY,
+    AUTOLOOT,
+    AUTOBANK,
+    RATE_USE_STAGES,
+    INVENTORY_GLOW,
+    TELEPORT_SUMMONS,
+    TOGGLE_DOWNLOAD_MAP,
+    USE_ANY_DATAPACK_FOLDER,
+    ALLOW_RELOAD,
+    BOOSTED_BOSS_SLOT,
+    XP_DISPLAY_MODE,
+    TOGGLE_GOLD_POUCH_ALLOW_ANYTHING,
+    TOGGLE_GOLD_POUCH_QUICKLOOT_ONLY,
+    TOGGLE_SERVER_IS_RETRO,
+    TOGGLE_TRAVELS_FREE,
+    TELEPORT_PLAYER_TO_VOCATION_ROOM,
+    OLD_PROTOCOL,
+    TOGGLE_HAZARDSYSTEM,
+    LOYALTY_ENABLED,
+    PARTY_AUTO_SHARE_EXPERIENCE,
+    PARTY_SHARE_LOOT_BOOSTS,
+    RESET_SESSIONS_ON_STARTUP,
+    TOGGLE_WHEELSYSTEM,
+    TOGGLE_ATTACK_SPEED_ONFIST,
+    VIP_SYSTEM_ENABLED,
+    VIP_AUTOLOOT_VIP_ONLY,
+    VIP_STAY_ONLINE,
+    REWARD_CHEST_COLLECT_ENABLED,
+    TOGGLE_MOUNT_IN_PZ,
+    TOGGLE_HOUSE_TRANSFER_ON_SERVER_RESTART,
+    TOGGLE_RECEIVE_REWARD,
+    TOGGLE_MAINTAIN_MODE,
+
+    LAST_BOOLEAN_CONFIG
+}
