+++ src/Extensions/NeoServer.Scripts.LuaJIT/Enums/Config/FloatingConfigType.cs
@@ -1,0 +1,39 @@
+namespace NeoServer.Scripts.LuaJIT.Enums.Config;
+
+public enum FloatingConfigType
+{
+    BESTIARY_RATE_CHARM_SHOP_PRICE,
+
+    RATE_HEALTH_REGEN,
+    RATE_HEALTH_REGEN_SPEED,
+    RATE_MANA_REGEN,
+    RATE_MANA_REGEN_SPEED,
+    RATE_SOUL_REGEN,
+    RATE_SOUL_REGEN_SPEED,
+
+    RATE_SPELL_COOLDOWN,
+    RATE_ATTACK_SPEED,
+    RATE_OFFLINE_TRAINING_SPEED,
+    RATE_EXERCISE_TRAINING_SPEED,
+
+    RATE_MONSTER_HEALTH,
+    RATE_MONSTER_ATTACK,
+    RATE_MONSTER_DEFENSE,
+    RATE_BOSS_HEALTH,
+    RATE_BOSS_ATTACK,
+    RATE_BOSS_DEFENSE,
+
+    RATE_NPC_HEALTH,
+    RATE_NPC_ATTACK,
+    RATE_NPC_DEFENSE,
+    LOYALTY_BONUS_PERCENTAGE_MULTIPLIER,
+    PARTY_SHARE_LOOT_BOOSTS_DIMINISHING_FACTOR,
+
+    PVP_RATE_DAMAGE_TAKEN_PER_LEVEL,
+    PVP_RATE_DAMAGE_REDUCTION_PER_LEVEL,
+
+    HOUSE_PRICE_RENT_MULTIPLIER,
+    HOUSE_RENT_RATE,
+
+    LAST_FLOATING_CONFIG
+}
