+++ src/Extensions/NeoServer.Scripts.LuaJIT/Enums/Config/IntegerConfigType.cs
@@ -1,0 +1,134 @@
+namespace NeoServer.Scripts.LuaJIT.Enums.Config;
+
+public enum IntegerConfigType
+{
+    SQL_PORT,
+    MAX_PLAYERS,
+    PZ_LOCKED,
+    DEFAULT_DESPAWNRANGE,
+    DEFAULT_DESPAWNRADIUS,
+    RATE_EXPERIENCE,
+    RATE_SKILL,
+    RATE_LOOT,
+    RATE_MAGIC,
+    RATE_SPAWN,
+    RATE_KILLING_IN_THE_NAME_OF_POINTS,
+    HOUSE_PRICE_PER_SQM,
+    HOUSE_BUY_LEVEL,
+    MAX_MESSAGEBUFFER,
+    ACTIONS_DELAY_INTERVAL,
+    EX_ACTIONS_DELAY_INTERVAL,
+    KICK_AFTER_MINUTES,
+    PROTECTION_LEVEL,
+    DEATH_LOSE_PERCENT,
+    STATUSQUERY_TIMEOUT,
+    FRAG_TIME,
+    WHITE_SKULL_TIME,
+    GAME_PORT,
+    LOGIN_PORT,
+    STATUS_PORT,
+    STAIRHOP_DELAY,
+    MAX_CONTAINER,
+    MAX_CONTAINER_ITEM,
+    MARKET_OFFER_DURATION,
+    DEPOT_BOXES,
+    FREE_DEPOT_LIMIT,
+    PREMIUM_DEPOT_LIMIT,
+    CHECK_EXPIRED_MARKET_OFFERS_EACH_MINUTES,
+    MAX_MARKET_OFFERS_AT_A_TIME_PER_PLAYER,
+    EXP_FROM_PLAYERS_LEVEL_RANGE,
+    MAX_PACKETS_PER_SECOND,
+    COMPRESSION_LEVEL,
+    STORE_COIN_PACKET,
+    DAY_KILLS_TO_RED,
+    WEEK_KILLS_TO_RED,
+    MONTH_KILLS_TO_RED,
+    RED_SKULL_DURATION,
+    BLACK_SKULL_DURATION,
+    ORANGE_SKULL_DURATION,
+    GLOBAL_SERVER_SAVE_NOTIFY_DURATION,
+    PUSH_DELAY,
+    PUSH_DISTANCE_DELAY,
+    STASH_ITEMS,
+    PARTY_LIST_MAX_DISTANCE,
+    STAMINA_ORANGE_DELAY,
+    STAMINA_GREEN_DELAY,
+    STAMINA_TRAINER_DELAY,
+    STAMINA_PZ_GAIN,
+    STAMINA_TRAINER_GAIN,
+    SAVE_INTERVAL_TIME,
+    PREY_REROLL_PRICE_LEVEL,
+    PREY_SELECTION_LIST_PRICE,
+    PREY_BONUS_TIME,
+    PREY_BONUS_REROLL_PRICE,
+    PREY_FREE_REROLL_TIME,
+    TASK_HUNTING_LIMIT_EXHAUST,
+    TASK_HUNTING_REROLL_PRICE_LEVEL,
+    TASK_HUNTING_SELECTION_LIST_PRICE,
+    TASK_HUNTING_BONUS_REROLL_PRICE,
+    TASK_HUNTING_FREE_REROLL_TIME,
+    MAX_ALLOWED_ON_A_DUMMY,
+    FREE_QUEST_STAGE,
+    DEPOTCHEST,
+    CRITICALCHANCE,
+    ADVENTURERSBLESSING_LEVEL,
+    FORGE_MAX_ITEM_TIER,
+    FORGE_COST_ONE_SLIVER,
+    FORGE_SLIVER_AMOUNT,
+    FORGE_CORE_COST,
+    FORGE_MAX_DUST,
+    FORGE_FUSION_DUST_COST,
+    FORGE_TRANSFER_DUST_COST,
+    FORGE_BASE_SUCCESS_RATE,
+    FORGE_BONUS_SUCCESS_RATE,
+    FORGE_TIER_LOSS_REDUCTION,
+    FORGE_AMOUNT_MULTIPLIER,
+    FORGE_MIN_SLIVERS,
+    FORGE_MAX_SLIVERS,
+    FORGE_INFLUENCED_CREATURES_LIMIT,
+    FORGE_FIENDISH_CREATURES_LIMIT,
+    BESTIARY_KILL_MULTIPLIER,
+    BOSSTIARY_KILL_MULTIPLIER,
+    BOOSTED_BOSS_LOOT_BONUS,
+    BOOSTED_BOSS_KILL_BONUS,
+    FAMILIAR_TIME,
+    BUY_AOL_COMMAND_FEE,
+    BUY_BLESS_COMMAND_FEE,
+    HAZARD_CRITICAL_INTERVAL,
+    HAZARD_CRITICAL_CHANCE,
+    HAZARD_CRITICAL_MULTIPLIER,
+    HAZARD_DAMAGE_MULTIPLIER,
+    HAZARD_DODGE_MULTIPLIER,
+    HAZARD_PODS_DROP_MULTIPLIER,
+    HAZARD_PODS_TIME_TO_DAMAGE,
+    HAZARD_EXP_BONUS_MULTIPLIER,
+    HAZARD_LOOT_BONUS_MULTIPLIER,
+    HAZARD_PODS_DAMAGE,
+    HAZARD_PODS_TIME_TO_SPAWN,
+    HAZARD_SPAWN_PLUNDER_MULTIPLIER,
+    LOW_LEVEL_BONUS_EXP,
+    LOYALTY_POINTS_PER_CREATION_DAY,
+    LOYALTY_POINTS_PER_PREMIUM_DAY_SPENT,
+    LOYALTY_POINTS_PER_PREMIUM_DAY_PURCHASED,
+    WHEEL_POINTS_PER_LEVEL,
+    MULTIPLIER_ATTACKONFIST,
+    MAX_SPEED_ATTACKONFIST,
+    TIBIADROME_CONCOCTION_COOLDOWN,
+    TIBIADROME_CONCOCTION_DURATION,
+    T_CONST,
+    PARALLELISM,
+    BOSS_DEFAULT_TIME_TO_FIGHT_AGAIN,
+    BOSS_DEFAULT_TIME_TO_DEFEAT,
+
+    VIP_BONUS_EXP,
+    VIP_BONUS_LOOT,
+    VIP_BONUS_SKILL,
+    VIP_FAMILIAR_TIME_COOLDOWN_REDUCTION,
+
+    REWARD_CHEST_MAX_COLLECT_ITEMS,
+    DISCORD_WEBHOOK_DELAY_MS,
+
+    PVP_MAX_LEVEL_DIFFERENCE,
+
+    LAST_INTEGER_CONFIG
+}
