+++ src/Extensions/NeoServer.Scripts.LuaJIT/Enums/Config/StringConfigType.cs
@@ -1,0 +1,39 @@
+namespace NeoServer.Scripts.LuaJIT.Enums.Config;
+
+public enum StringConfigType
+{
+    MAP_NAME,
+    MAP_DOWNLOAD_URL,
+    MAP_AUTHOR,
+    HOUSE_RENT_PERIOD,
+    SERVER_NAME,
+    SERVER_MOTD,
+    OWNER_NAME,
+    OWNER_EMAIL,
+    URL,
+    LOCATION,
+    IP,
+    WORLD_TYPE,
+    MYSQL_HOST,
+    MYSQL_USER,
+    MYSQL_PASS,
+    MYSQL_DB,
+    MYSQL_SOCK,
+    AUTH_TYPE,
+    DEFAULT_PRIORITY,
+    STORE_IMAGES_URL,
+    MAP_CUSTOM_NAME,
+    MAP_CUSTOM_AUTHOR,
+    DISCORD_WEBHOOK_URL,
+    SAVE_INTERVAL_TYPE,
+    GLOBAL_SERVER_SAVE_TIME,
+    DATA_DIRECTORY,
+    CORE_DIRECTORY,
+    FORGE_FIENDISH_INTERVAL_TYPE,
+    FORGE_FIENDISH_INTERVAL_TIME,
+    TIBIADROME_CONCOCTION_TICK_TYPE,
+    M_CONST,
+    MAINTAIN_MODE_MESSAGE,
+
+    LAST_STRING_CONFIG
+}
