+++ src/Extensions/NeoServer.Scripts.LuaJIT/Functions/ConfigFunctions.cs
@@ -1,6 +1,7 @@
 ﻿using LuaNET;
 using Microsoft.Extensions.Configuration;
 using NeoServer.Scripts.LuaJIT.Enums;
+using NeoServer.Scripts.LuaJIT.Enums.Config;
 using NeoServer.Scripts.LuaJIT.Functions.Interfaces;
 using NeoServer.Scripts.LuaJIT.Interfaces;
 using NeoServer.Server.Configurations;
