+++ src/Extensions/NeoServer.Scripts.LuaJIT/Interfaces/IConfigManager.cs
@@ -1,5 +1,6 @@
 ﻿using LuaNET;
 using NeoServer.Scripts.LuaJIT.Enums;
+using NeoServer.Scripts.LuaJIT.Enums.Config;
 
 namespace NeoServer.Scripts.LuaJIT.Interfaces;
