+++ src/Extensions/NeoServer.Scripts.LuaJIT/Scripts.cs
@@ -1,6 +1,7 @@
 ﻿using NeoServer.Domain.Common.Contracts.DataStores;
 using NeoServer.Domain.Spells;
 using NeoServer.Scripts.LuaJIT.Enums;
+using NeoServer.Scripts.LuaJIT.Enums.Config;
 using NeoServer.Scripts.LuaJIT.Interfaces;
 using Serilog;
